Working display and switching images, before refactoring to break glHandle to multiple classes

+++ ToneMapper/textureFragmentShader.glsl
@@ -1,7 +1,7 @@
 
-varying lowp vec2 TexCoordOut; // New
-uniform sampler2D Texture; // New
+varying highp vec2 TexCoordOut;
+uniform sampler2D Texture;
 
 void main(void) {
-  gl_FragColor = texture2D(Texture, TexCoordOut); // New
+  gl_FragColor = texture2D(Texture, TexCoordOut);
 }
